output - ROI and SHIFT algorithms
The number of the collected vegetable images was used to develop and train the system; Area and perimeter values for each type of vegetable that have been stored during system training. These stored shape roundness values, area and perimeter values are being used as standard feature values for comparison and classification of query or input vegetable image to the system. 
The images of the vegetable are uploaded to Raspberry-Pi. Now based on image processing the images are analysed and detect the defected area in row and column of the pixel. After detecting any defected area, it will check the neighborhood pixel of row and column and then match the trained input. Finally, the defect will be identified. 
Separation is done based on the total number of keypoints. Rotten goods have keypoints more than the fresh goods because of the shrinkages and cracks on the surface. We provide a threshold value so that if the keypoint value is more than the threshold, the servo motor allows a 45-degree left deviation and vice-versa.