The Whole Project Setup
Conveyor setup is made up of a transparent belt carrying the tomatoes in a linear way by a gear system. When the tomatoes are made to run, it is stopped for a specific time and the camera captures the tomatoes. This image is processed via Matlab through Raspberrypi. Now after processing by necessary algorithms, a servo motor is used to separate the good and rotten tomatoes by a separator system.